feat: Release activity library for everyone
