+++ packages/server/graphql/public/mutations/signUpWithPassword.ts
@@ -65,7 +65,7 @@ const signUpWithPassword: MutationResolvers['signUpWithPassword'] = async (
   const hashedPassword = await bcrypt.hash(password, Security.SALT_ROUNDS)
   const newUser = createNewLocalUser({email, hashedPassword, isEmailVerified: false, pseudoId})
   // MUTATIVE
-  context.authToken = await bootstrapNewUser(newUser, isOrganic, dataLoader, params)
+  context.authToken = await bootstrapNewUser(newUser, isOrganic, dataLoader)
   return {
     userId: newUser.id,
     authToken: encodeAuthToken(context.authToken),
