+++ packages/server/graphql/public/types/UserFeatureFlags.ts
@@ -4,7 +4,6 @@ const UserFeatureFlags: UserFeatureFlagsResolvers = {
   insights: ({insights}) => !!insights,
   noAISummary: ({noAISummary}) => !!noAISummary,
   noMeetingHistoryLimit: ({noMeetingHistoryLimit}) => !!noMeetingHistoryLimit,
-  retrosInDisguise: ({retrosInDisguise}) => !!retrosInDisguise,
   adHocTeams: ({adHocTeams}) => !!adHocTeams,
   noTemplateLimit: ({noTemplateLimit}) => !!noTemplateLimit,
   signUpDestinationTeam: ({signUpDestinationTeam}) => !!signUpDestinationTeam
